requirements.txt
flask
pandas
google-genai
fredapi
python-dotenv
gunicorn